chore(com): Tweak tests and naming
