+++ communication/src/main/java/datadog/communication/http/client/HttpClient.java
@@ -3,7 +3,7 @@ package datadog.communication.http.client;
 import java.io.IOException;
 
 /** Facade for synchronous HTTP calls independent from the underlying client implementation. */
-public interface HttpClientFacade extends AutoCloseable {
+public interface HttpClient extends AutoCloseable {
 
   HttpClientResponse execute(HttpClientRequest request) throws IOException;
