+++ communication/src/main/java/datadog/communication/http/client/HttpClientFactory.java
@@ -3,5 +3,5 @@ package datadog.communication.http.client;
 /** Factory for creating transport-specific HTTP client facades. */
 public interface HttpClientFactory {
 
-  HttpClientFacade create();
+  HttpClient create();
 }
